# Update exp package