--- a/cmd/promtool/tsdb.go
+++ b/cmd/promtool/tsdb.go
@@ -446,7 +446,7 @@ func analyzeBlock(path, blockID string, limit int, runExtended bool) error {
 	postingInfos := []postingInfo{}
 
 	printInfo := func(postingInfos []postingInfo) {
-		slices.SortFunc(postingInfos, func(a, b postingInfo) bool { return a.metric > b.metric })
+		slices.SortFunc(postingInfos, func(a, b postingInfo) int { return int(b.metric - a.metric) })
 
 		for i, pc := range postingInfos {
 			if i >= limit {