--- a/storage/remote/codec.go
+++ b/storage/remote/codec.go
@@ -186,8 +186,8 @@ func FromQueryResult(sortSeries bool, res *prompb.QueryResult) storage.SeriesSet
 	}
 
 	if sortSeries {
-		slices.SortFunc(series, func(a, b storage.Series) bool {
-			return labels.Compare(a.Labels(), b.Labels()) < 0
+		slices.SortFunc(series, func(a, b storage.Series) int {
+			return labels.Compare(a.Labels(), b.Labels())
 		})
 	}
 	return &concreteSeriesSet{