--- a/storage/remote/read_handler.go
+++ b/storage/remote/read_handler.go
@@ -16,6 +16,7 @@ package remote
 import (
 	"context"
 	"net/http"
+	"strings"
 	"sync"
 
 	"github.com/go-kit/log"
@@ -92,8 +93,8 @@ func (h *readHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 			Value: value,
 		})
 	}
-	slices.SortFunc(sortedExternalLabels, func(a, b prompb.Label) bool {
-		return a.Name < b.Name
+	slices.SortFunc(sortedExternalLabels, func(a, b prompb.Label) int {
+		return strings.Compare(a.Name, b.Name)
 	})
 
 	responseType, err := NegotiateResponseType(req.AcceptedResponseTypes)