--- a/tsdb/exemplar.go
+++ b/tsdb/exemplar.go
@@ -185,8 +185,8 @@ func (ce *CircularExemplarStorage) Select(start, end int64, matchers ...[]*label
 		}
 	}
 
-	slices.SortFunc(ret, func(a, b exemplar.QueryResult) bool {
-		return labels.Compare(a.SeriesLabels, b.SeriesLabels) < 0
+	slices.SortFunc(ret, func(a, b exemplar.QueryResult) int {
+		return labels.Compare(a.SeriesLabels, b.SeriesLabels)
 	})
 
 	return ret, nil