--- a/tsdb/wlog/checkpoint.go
+++ b/tsdb/wlog/checkpoint.go
@@ -374,8 +374,8 @@ func listCheckpoints(dir string) (refs []checkpointRef, err error) {
 		refs = append(refs, checkpointRef{name: fi.Name(), index: idx})
 	}
 
-	slices.SortFunc(refs, func(a, b checkpointRef) bool {
-		return a.index < b.index
+	slices.SortFunc(refs, func(a, b checkpointRef) int {
+		return a.index - b.index
 	})
 
 	return refs, nil