imo, this is worse in term or readability :(

Fixes https://github.com/prometheus/prometheus/issues/12649