@gouthamve 2023-08-04
So the test failure is interesting. Casting a float to an int leads to rounding to `0` which means equivalence in the compare function.

How to handle it in a simple manner is eluding me. Thoughts anyone? cc @bboreham @roidelapluie  

In `promql/quantile.go`:

```diff
-	slices.SortFunc(buckets, func(a, b bucket) bool {
-		return a.upperBound < b.upperBound
+	slices.SortFunc(buckets, func(a, b bucket) int {
+		return int(a.upperBound - b.upperBound)
```


https://play.golang.com/p/M03btR83odb