@gouthamve 2023-08-04
There is https://pkg.go.dev/cmp#Compare but it's not released yet and bumps the requirement to Golang 1.21