@bboreham 2023-08-05
I think we should benchmark this; the extra work required to produce ternary results is relatively high.  Also `strings.Compare` is deliberately not optimized, for reasons that are now invalid.

If it turns out this hurts performance, noting that the implementation of `SortFunc` only ever uses the function as `cmp(...) < 0`, we could cheat and only implement that side.

Maybe Go 1.21 has other improvements that regain the lost performance. Let's measure, anyway.

Regarding the float comparison, it seems the Go team went for the complicated version: https://github.com/golang/exp/blob/d63ba01acd4b5892385c4390b334ea2c89f27eb7/slices/cmp.go#L30-L44