@bboreham 2023-08-15
For at least one benchmark, which is dominated by the sorting time, this change made no difference:

```
cpu: Intel(R) Core(TM) i5-9400F CPU @ 2.90GHz
          │  before.txt  │          after.txt           │
          │    sec/op    │   sec/op     vs base         │
Builder-4   784.0n ± 30%   781.8n ± 4%  ~ (p=1.000 n=6)

          │ before.txt │           after.txt           │
          │    B/op    │    B/op     vs base           │
Builder-4   288.0 ± 0%   288.0 ± 0%  ~ (p=1.000 n=6) ¹
¹ all samples are equal

          │ before.txt │           after.txt           │
          │ allocs/op  │ allocs/op   vs base           │
Builder-4   1.000 ± 0%   1.000 ± 0%  ~ (p=1.000 n=6) ¹
¹ all samples are equal
```